requirements.txt
Flask==3.0.0
pandas==2.2.3
openpyxl==3.1.2
gunicorn==21.2.0
Werkzeug==3.0.0
python-dotenv==1.0.0 